AddBot message sending

+++ BarsantiExplorer/Models/Responses/UserResponse.cs
@@ -3,5 +3,5 @@ namespace BarsantiExplorer.Models.Responses;
 public class UserResponse
 {
     public string Email { get; set; }
-    public int? TelegramId { get; set; }
+    public long? TelegramId { get; set; }
 }
